db/mysql_connection.py
import os
import mysql.connector
from dotenv import load_dotenv
load_dotenv()

def conectar_mysql():
    return mysql.connector.connect(
        host=os.getenv("MYSQL_HOST"),
        port=int(os.getenv("MYSQL_PORT", 3306)),
        user=os.getenv("MYSQL_USER"),
        password=os.getenv("MYSQL_PASSWORD"),
        database=os.getenv("MYSQL_DB"),
        connection_timeout=10,
    )
