requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
mysql-connector-python>=8.3.0
python-dotenv>=1.0.1
openpyxl>=3.1.2
plotly>=5.19.0
streamlit-option-menu>=0.3.12
streamlit-extras>=0.4.2
xlsxwriter>=3.2.0
